+++ src/main/java/roidrole/roidtweaker/Utils.java
@@ -1,0 +1,7 @@
+package roidrole.roidtweaker;
+
+public class Utils {
+    public static int ceilDiv(int x, int y){
+        return -Math.floorDiv(-x,y);
+    }
+}